+++ dal/dao.py
@@ -1,5 +1,4 @@
 import sqlite3
-import os
 
 
 class Database:
