+++ hawqal/states.py
@@ -1,19 +1,19 @@
-from hawqal.dal.dao import Database
+from dal.dao import Database
 import string
 
 
-class CitiesByCountry:
+class StatesByCountry:
 
     @staticmethod
-    def getCities(country):
+    def getStates(country):
         states = []
         country = string.capwords(country)
-        file_name = "database/hawqalDB.sqlite"
+        file_name = "/hawqal/database/hawqalDB.sqlite"
         with open(file_name, 'r', encoding="utf8") as db:
             database = Database(file_name).makeConnection()
             cursor = database.cursor()
             data = cursor.execute(
-                f'SELECT name FROM cities WHERE country_name = "{country}"')
+                f'SELECT name FROM states WHERE country_name = "{country}"')
             for row in data:
                 states.append(f'{row[0]}')
         return states
