+++ setup.py
@@ -1,46 +1,16 @@
 """Python setup.py for hawqal package"""
-import io
-import os
 from setuptools import find_packages, setup
-
-
-def read(*paths, **kwargs):
-    """Read the contents of a text file safely.
-    >>> read("hawqal", "VERSION")
-    '0.1.0'
-    >>> read("README.md")
-    ...
-    """
-
-    content = ""
-    with io.open(
-        os.path.join(os.path.dirname(__file__), *paths),
-        encoding=kwargs.get("encoding", "utf8"),
-    ) as open_file:
-        content = open_file.read().strip()
-    return content
-
-
-def read_requirements(path):
-    return [
-        line.strip()
-        for line in read(path).split("\n")
-        if not line.startswith(('"', "#", "-", "git+"))
-    ]
 
 
 setup(
     name="hawqal",
-    version=read("hawqal", "0.1.0"),
+    version="0.1.0",
     description="Python package that contains the data of world's countries,states and their cities name",
     url="https://github.com/CapregSoft/Hawqal-python.git",
-    long_description=read("README.md"),
     long_description_content_type="text/markdown",
     author="capregsoft",
     packages=find_packages(exclude=["tests", ".github"]),
-    install_requires=read_requirements("requirements.txt"),
     entry_points={
         "console_scripts": ["hawqal = hawqal.__main__:main"]
     },
-    extras_require={"test": read_requirements("requirements.txt")},
 )
